requirements.txt
accelerate==1.8.1
aenum==2.2.6
aiofiles==24.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiohttp-retry==2.8.3
aioredis==2.0.1
aiosignal==1.3.1
aiosqlite==0.21.0
altair==5.3.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
arrow==1.3.0
asgiref==3.8.1
assemblyai==0.26.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
Authlib==1.3.1
autobahn==23.6.2
Automat==22.10.0
azure-cognitiveservices-speech==1.37.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.7.0
boto==2.49.0
boto3==1.35.0
botocore==1.35.0
build==1.2.2
bunq==0.0.2
click==8.2.1
cloc==0.2.5
colorama==0.4.6
colored==2.2.4
coloredlogs==15.0.1
constantly==23.10.4
contourpy==1.2.1
crontab==1.0.1
crypto==1.4.1
cryptography==45.0.2
cssselect==1.3.0
cycler==0.12.1
daily-python==0.10.1
daphne==4.1.2
dataclasses-json==0.5.14
decorator==4.4.2
Deprecated==1.2.14
deprecation==2.1.0
distlib==0.3.8
distro==1.9.0
dj-database-url==3.0.1
dj-stripe==2.8.4
Django==5.0.4
django-admin-two-factor==0.1.1
django-appconf==1.0.6
django-cookie-consent==0.4.0
django-cors-headers==4.3.1
django-cryptography==1.1
django-elasticsearch-dsl==9.0
django-elasticsearch-dsl-drf==0.22.5
django-encrypted-model-fields==0.6.5
django-environ==0.11.2
django-honeypot==1.1.0
django-mfa==3.2
django-nine==0.2.7
django-ninja==1.1.0
django-otp==0.3.14
django-phonenumber-field==8.0.0
django-rq==3.0.0
django-storages==1.14.3
django-turnstile==0.1.0
django-utils-six==2.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
docutils==0.21.2
dotenv==0.9.9
drf-spectacular==0.28.0
drf-yasg==1.21.7
durationpy==0.8
effdet==0.4.1
elastic-transport==8.17.1
elasticsearch==9.0.2
elasticsearch-dsl==0.0.9
elementpath==4.4.0
elevenlabs==1.2.1
email_validator==2.2.0
emoji==2.14.1
environ==1.0
executing==2.0.1
fastapi==0.115.12
fastapi-cli==0.0.2
feedparser==6.0.11
filelock==3.15.4
filetype==1.2.0
Flask==3.0.3
flask-sock==0.7.0
flatbuffers==24.3.25
fonttools==4.51.0
freezegun==1.5.1
frozenlist==1.4.1
fsspec==2024.9.0
fuzzywuzzy==0.18.0
gevent==24.2.1
gitdb==4.0.11
GitPython==3.1.43
google==3.0.0
google-api-core==2.25.1
google-api-python-client==2.174.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-vision==3.10.2
google-play-scraper==1.2.7
googleapis-common-protos==1.65.0
gotrue==2.12.3
greenlet==3.2.2
grpcio==1.73.1
grpcio-status==1.71.0
gspread==6.1.0
h11==0.14.0
h2==4.2.0
h3==4.2.2
hf-xet==1.1.5
hpack==4.1.0
html5lib==1.1
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.0
httpx-sse==0.4.0
hubspot-api-client==11.1.0
huggingface-hub==0.33.0
humanfriendly==10.0
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.10
imageio==2.34.1
imageio-ffmpeg==0.4.9
importlib-metadata==7.0.0
importlib_resources==6.4.5
incremental==22.10.0
inflection==0.5.1
iniconfig==2.1.0
ipython==8.24.0
itsdangerous==2.2.0
janus==1.0.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.0.2
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
joblib==1.4.2
jq==1.9.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
july==0.1.3
keyring==25.3.0
kiwisolver==1.4.5
kubernetes==31.0.0
langchain==0.3.26
langchain-chroma==0.2.4
langchain-community==0.3.26
langchain-core==0.3.66
langchain-huggingface==0.3.0
langchain-ollama==0.3.3
langchain-openai==0.3.24
langchain-text-splitters==0.3.8
langdetect==1.0.9
langsmith==0.4.1
Levenshtein==0.26.0
libmagic==1.0
lightsaillogger==0.1.6
lxml==5.4.0
lxml_html_clean==0.4.2
mangum==0.17.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib==3.7.0
matplotlib-inline==0.1.7
mcp==1.7.1
mdurl==0.1.2
mixpanel==4.10.1
mmh3==5.0.1
monotonic==1.6
more-itertools==10.4.0
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
mysqlclient==2.2.6
Naked==0.1.32
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.3
nh3==0.2.18
nltk==3.8.1
numexpr==2.10.0
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
ollama==0.5.1
omegaconf==2.3.0
onnx==1.18.0
onnxruntime==1.19.2
openai==1.86.0
openapi-schema-pydantic==1.2.4
opencv-python==4.11.0.86
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
orjson==3.10.2
overrides==7.7.0
packaging==24.0
pandas==2.2.2
parso==0.8.4
pdf2image==1.17.0
pdfminer.six==20250506
pexpect==4.9.0
phonenumbers==8.13.40
pi_heif==0.22.0
pikepdf==9.9.0
pillow==10.4.0
pkginfo==1.10.0
platformdirs==4.2.2
playwright==1.49.1
plotly==5.21.0
pluggy==1.6.0
postgrest==1.1.1
posthog==3.6.6
proglog==0.1.10
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.21.1
prompt-toolkit==3.0.43
proto-plus==1.26.1
protobuf==4.25.8
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAudio==0.2.14
pybase64==1.4.1
pycocotools==2.0.10
pycparser==2.22
pycrypto==2.6.1
pycryptodomex==3.21.0
pydantic==2.11.7
pydantic-settings==2.5.2
pydantic_core==2.33.2
pydeck==0.8.1b0
pydub==0.25.1
pyee==12.0.0
Pygments==2.18.0
PyJWT==2.10.1
PyMySQL==1.1.1
pyngrok==7.1.6
pyOpenSSL==24.1.0
pyotp==2.9.0
pyparsing==3.1.2
pypdf==5.6.1
pypdfium2==4.30.1
PyPika==0.48.9
pypng==0.20220715.0
pyproject_hooks==1.2.0
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-gitlab==4.4.0
python-iso639==2025.2.18
python-Levenshtein==0.26.0
python-magic==0.4.27
python-multipart==0.0.9
python-oxmsg==0.0.2
python-telegram-bot==22.0
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
RapidFuzz==3.10.0
rdflib==7.1.4
readability==0.3.2
readability-lxml==0.8.4.1
readme_renderer==44.0
realtime==2.6.0
redis==5.0.8
Redis-Sentinel-Url==1.0.1
referencing==0.34.0
regex==2024.4.28
reportlab==4.2.5
requests==2.32.3
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.1
rpds-py==0.18.0
rq==2.1.0
rq-dashboard==0.8.2.2
rq-dashboard-fast==0.5.8
rq-gevent-worker==0.1.6
rq-scheduler==0.14.0
rsa==4.9
ruff==0.5.1
s3transfer==0.10.2
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
sentence-transformers==3.1.1
sentry-sdk==1.45.0
service-identity==24.1.0
setuptools==73.0.1
sgmllib3k==1.0.0
shellescape==3.8.1
shellingham==1.5.4
simple-websocket==1.0.0
simplejson==3.19.3
six==1.16.0
slack_sdk==3.27.1
smmap==5.0.1
sniffio==1.3.1
sounddevice==0.4.6
soupsieve==2.5
SQLAlchemy==2.0.41
sqlparse==0.4.4
sse-starlette==2.3.4
stack-data==0.6.3
starlette==0.46.2
StrEnum==0.4.15
stripe==10.8.0
supabase==2.17.0
whitenoise==6.6.0
